Execute code block even if Bolt is locked
You probably are wondering why is the purpose of this. The main reason
to execute a "critical section" without acquiring the mutex (Bolt) is
because we only need that the lock has been acquired so we know for sure that Ghost Tables are not going to be executed.

In which cases are we going to need this? Mainly in the register part of
the importer where we're going to perform DDL operations that are going
to trigger ghost tables jobs and that could coincide at the same time
with other DDL operations leading into a race condition.