Retry bolt block of code if needed
We've added the possibility to retry critical sections of code if Bolt
received an attempt to lock and that block of code is marked as
retriable.

Why are we doing this? Because we don't want to miss changes in the
user's schema because Bolt was already acquired.

For example, if some user is doing intensive DDL operations we want to
sync all the changes so to achieve that we need to inform Bolt that
we have other processes trying to execute Ghost tables at the same
time and instead of releasing the lock, we mark it asking for retry
and at the end we re-execute the same critical section again.